Refactor: always parse price as double in ShopCreationGUI, and minor code warning cleanups in Slabbo.java. Ensures price input logic is robust and consistent.
